Sithafal Task - 5/context-hub-evaluation/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0
stripe
# subprocess is part of Python's standard library and is used directly in the code.
